RegML should refer to schema on github.com
We don't want RegML files to refer to the schema on cfpb.github.io
because that requires keeping the schema as part of the Github Pages
documentation site. We'd rather refer to the schema as it exists in
source control, so this change modifies RegML to refer to github.com
instead.

Note that this refers to the `master` branch; potentially it would be
nice to refer to a tagged version of the RegML schema that updates each
time the RegML is (re)generated.